+++ spec/fixtures/dummy/app/controllers/api/activity_logs_controller.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Api
+  class ActivityLogsController < ApplicationController
+    # List activity logs grouped by recency.
+    def index; end
+  end
+end